Renamed the django application for clarity
