+++ django/README.md
@@ -5,7 +5,7 @@ This is a Django server for doing web application work.
 ## Overview
 
 - __Docker__ : recipe for building the django image
-- __kafka_test__ : example application
+- __kafka_example__ : example application
 - __www__ : project definition
   - __www/settings.py__ : server settings
   - __www/urls.py__ : top-level routing table
