+++ django/kafka_example/admin.py
