+++ django/kafka_example/channels/routing.py
@@ -1,7 +1,7 @@
 # chat/routing.py
 from django.urls import re_path
 
-from kafka_test.channels import consumers
+from kafka_example.channels import consumers
 
 websocket_urlpatterns = [
     re_path(r'example/ws/$', consumers.ExampleConsumer.as_asgi()),
