+++ django/kafka_example/models.py
