+++ django/kafka_example/static/kafka_example/kafka-example.css
