+++ django/kafka_example/templates/kafka_example/index.html
@@ -4,7 +4,7 @@
 <html>
   <head>
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
-    <link rel="stylesheet" href="{% static 'kafka_test/kafka-test.css' %}">
+    <link rel="stylesheet" href="{% static 'kafka_example/kafka-example.css' %}">
   </head>
   <body>
