+++ django/kafka_example/tests.py
