--- django/kafka_test/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-from kafka_test import views
-
-
-urlpatterns = [
-    path('', views.IndexView.as_view(), name='kafka-test'),
-]
